requirements.txt
Django~=4.0.1
psycopg2>=2.8
celery
redis==4.3.1
djangorestframework~=3.11.0
django-redis
django-celery-beat
pyyaml
uritemplate
django-rest-swagger
djangorestframework-simplejwt